Refactor ProtoOS firmware E2E helpers
